DHP-1079: manage yearFormat to allowValue mapping in YearRadioGroup
