+++ src/components/surveys/survey-design/left-panel/QuestionConfigs.tsx
@@ -13,7 +13,7 @@ import MultiSelect from '@mui/icons-material/DnsTwoTone'
 import SingleSelect from '@mui/icons-material/ToggleOffTwoTone'
 import {ChoiceQuestion, Question, Step} from '@typedefs/surveys'
 import React from 'react'
-import {DEFAULT_MIN_YEAR} from '../question-edit/rhs-subcontrols/Year'
+import {DEFAULT_MIN_YEAR} from '../question-edit/rhs-subcontrols/YearRadioGroup'
 
 export type QuestionTypeKey =
   | 'DURATION'
