+++ src/components/surveys/survey-design/question-edit/rhs-subcontrols/YearRadioGroup.tsx
@@ -3,23 +3,48 @@ import {Box, FormControl, FormControlLabel, Radio, RadioGroup} from '@mui/materi
 import {theme} from '@style/theme'
 import React from 'react'
 
+export const DEFAULT_MIN_YEAR = 1900
+export const DEFAULT_MAX_YEAR = new Date().getFullYear()
+
 export type YearFormatType = 'ANY' | 'CURRENT' | 'SET'
 
 const capitalizeFirstLetter = (val: string) => {
   return `${val.charAt(0).toUpperCase()}${val.substring(1)}`
 }
 
+const yearFormatToAllowValueMap: Record<YearFormatType, boolean | undefined> = {
+  SET: undefined,
+  ANY: true,
+  CURRENT: false,
+}
+
+function yearFormatToAllowValue(yearFormat: YearFormatType) {
+  return yearFormatToAllowValueMap[yearFormat]
+}
+
+function allowValueToYearFormat(allowValue: boolean | undefined): YearFormatType {
+  return (
+    (Object.keys(yearFormatToAllowValueMap) as YearFormatType[]).find(
+      key => yearFormatToAllowValueMap[key] === allowValue
+    ) || 'ANY'
+  )
+}
+
+function setYearValue(yearFormat: YearFormatType, yearValue: number | undefined) {
+  return yearFormat === 'SET' ? yearValue : undefined
+}
+
 type YearRadioGroupProps = {
   type: 'min' | 'max'
-  yearFormat: YearFormatType
+  allowValue: boolean | undefined
   yearValue: number | undefined
   hasError: boolean
-  onChange: (yearFormat: YearFormatType, yearValue: number | undefined) => void
+  onChange: (allowValue: boolean | undefined, yearValue: number | undefined) => void
 }
 
 const YearRadioGroup: React.FunctionComponent<YearRadioGroupProps> = ({
   type,
-  yearFormat,
+  allowValue,
   yearValue,
   hasError,
   onChange,
@@ -35,6 +60,8 @@ const YearRadioGroup: React.FunctionComponent<YearRadioGroupProps> = ({
       yearLabel: 'Maximum Year',
     },
   }
+  const yearFormat: YearFormatType = allowValueToYearFormat(allowValue)
+  const defaultYearValue = type === 'min' ? DEFAULT_MIN_YEAR : DEFAULT_MAX_YEAR
 
   const inlineInput = (
     <Box display="flex" alignItems="center">
@@ -47,7 +74,10 @@ const YearRadioGroup: React.FunctionComponent<YearRadioGroupProps> = ({
         type="number"
         disabled={yearValue === undefined}
         onChange={(e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => {
-          onChange(yearFormat, parseInt(e.target.value))
+          onChange(
+            yearFormatToAllowValue(yearFormat),
+            setYearValue(yearFormat, parseInt(e.target.value) || defaultYearValue)
+          )
         }}
       />
     </Box>
@@ -62,7 +92,8 @@ const YearRadioGroup: React.FunctionComponent<YearRadioGroupProps> = ({
           id={`${type}YearFormat`}
           value={yearFormat}
           onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-            onChange(e.target.value as YearFormatType, yearValue)
+            const newYearFormat = e.target.value as YearFormatType
+            onChange(yearFormatToAllowValue(newYearFormat), setYearValue(newYearFormat, yearValue || defaultYearValue))
           }}>
           <FormControlLabel
             value="ANY"
